+++ typescript/cli/src/deploy/warp.ts
@@ -3,28 +3,50 @@ import { stringify as yamlStringify } from 'yaml';
 
 import { IRegistry } from '@hyperlane-xyz/registry';
 import {
+  AggregationIsmConfig,
+  ChainName,
+  EvmERC20WarpModule,
   EvmIsmModule,
   HypERC20Deployer,
   HypERC721Deployer,
   HyperlaneAddresses,
   HyperlaneContractsMap,
   HyperlaneProxyFactoryDeployer,
+  IsmType,
   MultiProvider,
+  MultisigIsmConfig,
+  OpStackIsmConfig,
+  PausableIsmConfig,
+  ProxyFactoryFactoriesAddresses,
+  RoutingIsmConfig,
   TOKEN_TYPE_TO_STANDARD,
   TokenFactories,
-  TokenType,
+  TrustedRelayerIsmConfig,
   WarpCoreConfig,
   WarpRouteDeployConfig,
   getTokenConnectionId,
+  isCollateralConfig,
   isTokenMetadata,
   serializeContracts,
 } from '@hyperlane-xyz/sdk';
-import { ProtocolType, objMap, promiseObjAll } from '@hyperlane-xyz/utils';
+import {
+  ProtocolType,
+  assert,
+  objMap,
+  promiseObjAll,
+} from '@hyperlane-xyz/utils';
 
 import { readWarpRouteDeployConfig } from '../config/warp.js';
 import { MINIMUM_WARP_DEPLOY_GAS } from '../consts.js';
 import { WriteCommandContext } from '../context/types.js';
-import { log, logBlue, logGray, logGreen, logTable } from '../logger.js';
+import {
+  log,
+  logBlue,
+  logGray,
+  logGreen,
+  logRed,
+  logTable,
+} from '../logger.js';
 import {
   indentYamlOrJson,
   isFile,
@@ -39,7 +61,11 @@ import {
 
 interface DeployParams {
   context: WriteCommandContext;
-  configMap: WarpRouteDeployConfig;
+  warpDeployConfig: WarpRouteDeployConfig;
+}
+
+interface ApplyParams extends DeployParams {
+  warpCoreConfig: WarpCoreConfig;
 }
 
 export async function runWarpRouteDeploy({
@@ -74,10 +100,8 @@ export async function runWarpRouteDeploy({
 
   const deploymentParams = {
     context,
-    configMap: warpRouteConfig,
+    warpDeployConfig: warpRouteConfig,
   };
-
-  logBlue('Warp route deployment plan');
 
   await runDeployPlanStep(deploymentParams);
   const chains = Object.keys(warpRouteConfig);
@@ -97,15 +121,12 @@ export async function runWarpRouteDeploy({
   await completeDeploy(context, 'warp', initialBalances, userAddress, chains);
 }
 
-async function runDeployPlanStep({ context, configMap }: DeployParams) {
+async function runDeployPlanStep({ context, warpDeployConfig }: DeployParams) {
   const { skipConfirmation } = context;
 
-  logBlue('\nDeployment plan');
-  logGray('===============');
-  log(`Using token standard ${configMap.isNft ? 'ERC721' : 'ERC20'}`);
-  logTable(configMap);
+  displayWarpDeployPlan(warpDeployConfig);
 
-  if (skipConfirmation) return;
+  if (skipConfirmation || context.isDryRun) return;
 
   const isConfirmed = await confirm({
     message: 'Is this deployment plan correct?',
@@ -117,18 +138,18 @@ async function executeDeploy(params: DeployParams) {
   logBlue('All systems ready, captain! Beginning deployment...');
 
   const {
-    configMap,
+    warpDeployConfig,
     context: { registry, multiProvider, isDryRun, dryRunChain },
   } = params;
 
-  const deployer = configMap.isNft
+  const deployer = warpDeployConfig.isNft
     ? new HypERC721Deployer(multiProvider)
     : new HypERC20Deployer(multiProvider);
 
   const config: WarpRouteDeployConfig =
     isDryRun && dryRunChain
-      ? { [dryRunChain]: configMap[dryRunChain] }
-      : configMap;
+      ? { [dryRunChain]: warpDeployConfig[dryRunChain] }
+      : warpDeployConfig;
 
   const ismFactoryDeployer = new HyperlaneProxyFactoryDeployer(multiProvider);
 
@@ -251,7 +272,7 @@ async function createWarpIsm(
 }
 
 async function getWarpCoreConfig(
-  { configMap, context }: DeployParams,
+  { warpDeployConfig, context }: DeployParams,
   contracts: HyperlaneContractsMap<TokenFactories>,
 ): Promise<WarpCoreConfig> {
   const warpCoreConfig: WarpCoreConfig = { tokens: [] };
@@ -259,35 +280,31 @@ async function getWarpCoreConfig(
   // TODO: replace with warp read
   const tokenMetadata = await HypERC20Deployer.deriveTokenMetadata(
     context.multiProvider,
-    configMap,
+    warpDeployConfig,
   );
+  assert(
+    tokenMetadata && isTokenMetadata(tokenMetadata),
+    'Missing required token metadata',
+  );
+  const { decimals, symbol, name } = tokenMetadata;
+  assert(decimals, 'Missing decimals on token metadata');
 
   // First pass, create token configs
   for (const [chainName, contract] of Object.entries(contracts)) {
-    const config = configMap[chainName];
-    const metadata = {
-      ...tokenMetadata,
-      ...config,
-    };
+    const config = warpDeployConfig[chainName];
+    const collateralAddressOrDenom = isCollateralConfig(config)
+      ? config.token // gets set in the above deriveTokenMetadata()
+      : undefined;
 
-    if (!isTokenMetadata(metadata)) {
-      throw new Error('Missing required token metadata');
-    }
-
-    const { decimals } = metadata;
-    if (!decimals) {
-      throw new Error('Missing decimals on token metadata');
-    }
-
-    const collateralAddressOrDenom =
-      config.type === TokenType.collateral ? config.token : undefined;
     warpCoreConfig.tokens.push({
       chainName,
       standard: TOKEN_TYPE_TO_STANDARD[config.type],
-      ...metadata,
       decimals,
+      symbol,
+      name,
       addressOrDenom:
-        contract[configMap[chainName].type as keyof TokenFactories].address,
+        contract[warpDeployConfig[chainName].type as keyof TokenFactories]
+          .address,
       collateralAddressOrDenom,
     });
   }
@@ -314,4 +331,186 @@ async function getWarpCoreConfig(
   }
 
   return warpCoreConfig;
+}
+
+export async function runWarpRouteApply(params: ApplyParams) {
+  const {
+    warpDeployConfig,
+    warpCoreConfig,
+    context: { registry, multiProvider },
+  } = params;
+
+  // Addresses used to get static Ism factories
+  const addresses = await registry.getAddresses();
+
+  // Convert warpCoreConfig.tokens[] into a mapping of { [chainName]: Config }
+  // This allows O(1) reads within the loop
+  const warpCoreByChain = Object.fromEntries(
+    warpCoreConfig.tokens.map((token) => [token.chainName, token]),
+  );
+
+  // Attempt to update Warp Routes
+  // Can update existing or deploy new contracts
+  logGray(`Comparing target and onchain Warp configs`);
+  await promiseObjAll(
+    objMap(warpDeployConfig, async (chain, config) => {
+      try {
+        // Update Warp
+        config.ismFactoryAddresses = addresses[
+          chain
+        ] as ProxyFactoryFactoriesAddresses;
+        const evmERC20WarpModule = new EvmERC20WarpModule(multiProvider, {
+          config,
+          chain,
+          addresses: {
+            deployedTokenRoute: warpCoreByChain[chain].addressOrDenom!,
+          },
+        });
+        const transactions = await evmERC20WarpModule.update(config);
+
+        // Send Txs
+        if (transactions.length) {
+          for (const transaction of transactions) {
+            await multiProvider.sendTransaction(chain, transaction);
+          }
+
+          logGreen(`Warp config updated on ${chain}.`);
+        } else {
+          logGreen(
+            `Warp config on ${chain} is the same as target. No updates needed.`,
+          );
+        }
+      } catch (e) {
+        logRed(`Warp config on ${chain} failed to update.`, e);
+      }
+    }),
+  );
+}
+
+function displayWarpDeployPlan(deployConfig: WarpRouteDeployConfig) {
+  logBlue('\nWarp Route Deployment Plan');
+  logGray('==========================');
+  log(`📋 Token Standard: ${deployConfig.isNft ? 'ERC721' : 'ERC20'}`);
+
+  const { transformedDeployConfig, transformedIsmConfigs } =
+    transformDeployConfigForDisplay(deployConfig);
+
+  log('📋 Warp Route Config:');
+  logTable(transformedDeployConfig);
+  objMap(transformedIsmConfigs, (chain, ismConfigs) => {
+    log(`📋 ${chain} ISM Config(s):`);
+    ismConfigs.forEach((ismConfig) => {
+      logTable(ismConfig);
+    });
+  });
+}
+
+/* only used for transformIsmForDisplay type-sense */
+type IsmConfig =
+  | RoutingIsmConfig // type, owner, ownerOverrides, domain
+  | AggregationIsmConfig // type, modules, threshold
+  | MultisigIsmConfig // type, validators, threshold
+  | OpStackIsmConfig // type, origin, nativeBridge
+  | PausableIsmConfig // type, owner, paused, ownerOverrides
+  | TrustedRelayerIsmConfig; // type, relayer
+
+function transformDeployConfigForDisplay(deployConfig: WarpRouteDeployConfig) {
+  const transformedIsmConfigs: Record<ChainName, any[]> = {};
+  const transformedDeployConfig = objMap(deployConfig, (chain, config) => {
+    if (config.interchainSecurityModule)
+      transformedIsmConfigs[chain] = transformIsmConfigForDisplay(
+        config.interchainSecurityModule as IsmConfig,
+      );
+
+    return {
+      'NFT?': config.isNft ?? false,
+      Type: config.type,
+      Owner: config.owner,
+      Mailbox: config.mailbox,
+      'ISM Config(s)': config.interchainSecurityModule
+        ? 'See table(s) below.'
+        : 'No ISM config(s) specified.',
+    };
+  });
+
+  return {
+    transformedDeployConfig,
+    transformedIsmConfigs,
+  };
+}
+
+function transformIsmConfigForDisplay(ismConfig: IsmConfig): any[] {
+  const ismConfigs: any[] = [];
+  switch (ismConfig.type) {
+    case IsmType.AGGREGATION:
+      ismConfigs.push({
+        Type: ismConfig.type,
+        Threshold: ismConfig.threshold,
+        Modules: 'See table(s) below.',
+      });
+      ismConfig.modules.forEach((module) => {
+        ismConfigs.push(...transformIsmConfigForDisplay(module as IsmConfig));
+      });
+      return ismConfigs;
+    case IsmType.ROUTING:
+      return [
+        {
+          Type: ismConfig.type,
+          Owner: ismConfig.owner,
+          'Owner Overrides': ismConfig.ownerOverrides ?? 'Undefined',
+          Domains: 'See warp config for domain specification.',
+        },
+      ];
+    case IsmType.FALLBACK_ROUTING:
+      return [
+        {
+          Type: ismConfig.type,
+          Owner: ismConfig.owner,
+          'Owner Overrides': ismConfig.ownerOverrides ?? 'Undefined',
+          Domains: 'See warp config for domain specification.',
+        },
+      ];
+    case IsmType.MERKLE_ROOT_MULTISIG:
+      return [
+        {
+          Type: ismConfig.type,
+          Validators: ismConfig.validators,
+          Threshold: ismConfig.threshold,
+        },
+      ];
+    case IsmType.MESSAGE_ID_MULTISIG:
+      return [
+        {
+          Type: ismConfig.type,
+          Validators: ismConfig.validators,
+          Threshold: ismConfig.threshold,
+        },
+      ];
+    case IsmType.OP_STACK:
+      return [
+        {
+          Type: ismConfig.type,
+          Origin: ismConfig.origin,
+          'Native Bridge': ismConfig.nativeBridge,
+        },
+      ];
+    case IsmType.PAUSABLE:
+      return [
+        {
+          Type: ismConfig.type,
+          Owner: ismConfig.owner,
+          'Paused ?': ismConfig.paused,
+          'Owner Overrides': ismConfig.ownerOverrides ?? 'Undefined',
+        },
+      ];
+    case IsmType.TRUSTED_RELAYER:
+      return [
+        {
+          Type: ismConfig.type,
+          Relayer: ismConfig.relayer,
+        },
+      ];
+    default:
+      return [ismConfig];
+  }
 }