+++ typescript/infra/config/environments/mainnet3/supportedChainNames.ts
@@ -1,11 +1,12 @@
 // These chains may be any protocol type.
 // Placing them here instead of adjacent chains file to avoid circular dep
-export const supportedChainNames = [
+export const mainnet3SupportedChainNames = [
   'arbitrum',
   'ancient8',
   'avalanche',
   'base',
   'blast',
+  'bob',
   'bsc',
   'celo',
   'ethereum',
@@ -15,6 +16,7 @@ export const supportedChainNames = [
   'injective',
   'linea',
   'mantapacific',
+  'mantle',
   'mode',
   'moonbeam',
   'neutron',
@@ -25,6 +27,9 @@ export const supportedChainNames = [
   'redstone',
   'scroll',
   'sei',
+  'taiko',
   'viction',
   'zetachain',
-];
+] as const;
+
+export const supportedChainNames = [...mainnet3SupportedChainNames];