+++ typescript/infra/config/environments/testnet4/funding.ts
@@ -3,8 +3,11 @@ import { Role } from '../../../src/roles.js';
 import { Contexts } from '../../contexts.js';
 
 import { environment } from './chains.js';
+import { testnet4SupportedChainNames } from './supportedChainNames.js';
 
-export const keyFunderConfig: KeyFunderConfig = {
+export const keyFunderConfig: KeyFunderConfig<
+  typeof testnet4SupportedChainNames
+> = {
   docker: {
     repo: 'gcr.io/abacus-labs-dev/hyperlane-monorepo',
     tag: 'efa9025-20240605-091304',
@@ -26,18 +29,37 @@ export const keyFunderConfig: KeyFunderConfig = {
     alfajores: '5',
     bsctestnet: '5',
     fuji: '5',
-    plumetestnet: '0.2',
     holesky: '5',
+    plumetestnet: '0.2',
     // Funder boosts itself upto 5x balance on L2 before dispersing funds
     scrollsepolia: '1',
     sepolia: '5',
+    // no funding for solana
+    eclipsetestnet: '0',
+    solanatestnet: '0',
   },
   desiredKathyBalancePerChain: {
     alfajores: '1',
     bsctestnet: '1',
     fuji: '1',
+    holesky: '0',
     plumetestnet: '0.05',
     scrollsepolia: '1',
     sepolia: '1',
+    // no funding for solana
+    eclipsetestnet: '0',
+    solanatestnet: '0',
+  },
+  igpClaimThresholdPerChain: {
+    alfajores: '1',
+    bsctestnet: '1',
+    fuji: '1',
+    holesky: '1',
+    plumetestnet: '0.1',
+    scrollsepolia: '0.1',
+    sepolia: '1',
+    // no funding for solana
+    eclipsetestnet: '0',
+    solanatestnet: '0',
   },
 };