+++ typescript/sdk/src/providers/transactions/schemas.ts
@@ -10,6 +10,8 @@ export const PopulatedTransactionSchema = z.object({
   chainId: z.number(),
 });
 
+export const PopulatedTransactionsSchema = PopulatedTransactionSchema.array();
+
 export const CallDataSchema = z.object({
   to: ZHash,
   data: z.string(),